style: Componentization
